Split extension.ts into modular architecture
Reorganize codebase into focused modules:
- src/providers/NetCDFTreeProvider.ts: Tree view provider and item classes
- src/views/datasetHtmlView.ts: HTML view generation for datasets
- src/views/variableWebview.ts: Chart.js webview for variable preview
- src/python/inspector.ts: Python script execution and dependency checks
- src/utils/sampleSlice.ts: Shared utility function (removes duplication)

Update extension.ts to be entry point only (~90 lines, down from ~540)
Update tests to import from new module locations
Update CONTRIBUTING.md with new project structure
