+++ src/extension.ts
@@ -1,25 +1,23 @@
-import { execFile } from 'child_process';
-import * as path from 'path';
 import * as vscode from 'vscode';
-import {
-  NetCDFDataset,
-  NetCDFVariable,
-  NamedVariable,
-  StoredNetCDF,
-  InspectResult,
-  isInspectError,
-} from './types';
+import { NetCDFDataset, StoredNetCDF, isInspectError } from './types';
+import { inspectNetCDFWithPython, checkPythonDependencies } from './python/inspector';
+import { NetCDFTreeProvider } from './providers/NetCDFTreeProvider';
+import { showDatasetHtmlView } from './views/datasetHtmlView';
 
 /**
  * Called when your extension is activated.
  * @param context VS Code extension context
  */
-export function activate(context: vscode.ExtensionContext) {
-  // 1) Register "Open NetCDF File…", now accepting an optional URI
+export function activate(context: vscode.ExtensionContext): void {
+  // Register TreeDataProvider for NetCDF Explorer
+  const provider = new NetCDFTreeProvider(context);
+  vscode.window.registerTreeDataProvider('netcdfExplorer', provider);
+
+  // Register "Open NetCDF File…" command, accepting an optional URI
   const openCmd = vscode.commands.registerCommand('netcdf-viewer.openFile', async (uri?: vscode.Uri) => {
     let fileUri = uri;
 
-    // 2) If no URI was passed (user ran from Command Palette), prompt for one
+    // If no URI was passed (user ran from Command Palette), prompt for one
     if (!fileUri) {
       const [picked] =
         (await vscode.window.showOpenDialog({
@@ -31,6 +29,7 @@ export function activate(context: vscode.ExtensionContext) {
       }
       fileUri = picked;
     }
+
     try {
       const result = await inspectNetCDFWithPython(context, fileUri.fsPath);
       if (isInspectError(result)) {
@@ -53,7 +52,7 @@ export function activate(context: vscode.ExtensionContext) {
   });
   context.subscriptions.push(openCmd);
 
-  // Register a command to select Python environment using the Python extension
+  // Register command to select Python environment
   const selectPythonEnvCmd = vscode.commands.registerCommand('netcdf-viewer.selectPythonEnv', async () => {
     const picked = await vscode.window.showOpenDialog({
       canSelectMany: false,
@@ -82,11 +81,7 @@ export function activate(context: vscode.ExtensionContext) {
   });
   context.subscriptions.push(showHtmlCmd);
 
-  // 3) Register TreeDataProvider for NetCDF Explorer
-  const provider = new NetCDFTreeProvider(context);
-  vscode.window.registerTreeDataProvider('netcdfExplorer', provider);
-
-  // Check Python and dependencies
+  // Check Python and dependencies on startup
   const pythonPath = vscode.workspace.getConfiguration().get<string>('netcdfViewer.pythonPath', 'python');
   checkPythonDependencies(pythonPath).then((depError) => {
     if (depError) {
@@ -98,446 +93,6 @@ export function activate(context: vscode.ExtensionContext) {
 /**
  * Called when your extension is deactivated.
  */
-export function deactivate() {
+export function deactivate(): void {
   // No cleanup needed currently
-}
-
-/**
- * Opens a Webview panel to preview the selected NetCDF variable.
- */
-function showVariableWebview(context: vscode.ExtensionContext, variable: NamedVariable) {
-  const panel = vscode.window.createWebviewPanel(
-    'netcdfVarPreview',
-    `Preview: ${variable.name || variable.label || '?'}`,
-    vscode.ViewColumn.One,
-    {
-      enableScripts: true,
-      localResourceRoots: [vscode.Uri.joinPath(context.extensionUri, 'media')],
-    }
-  );
-
-  panel.webview.html = getWebviewContent(panel.webview, context, variable);
-
-  panel.webview.onDidReceiveMessage((msg) => {
-    if (msg.command === 'alert') {
-      vscode.window.showInformationMessage(msg.text);
-    }
-  });
-}
-
-/** Opens a webview to display the entire dataset as an HTML table with collapsible sections. */
-function showDatasetHtmlView(context: vscode.ExtensionContext, dataset: NetCDFDataset) {
-  const stored = context.workspaceState.get<StoredNetCDF>('lastNetCDF');
-  const fileName = stored && stored.uri ? require('path').basename(stored.uri.fsPath) : 'NetCDF File';
-  const panel = vscode.window.createWebviewPanel(
-    'netcdfHtmlView',
-    fileName, // Use the filename as the tab heading
-    vscode.ViewColumn.One,
-    { enableScripts: true }
-  );
-  panel.webview.html = getDatasetHtml(panel.webview, dataset, fileName);
-}
-
-/** Returns HTML markup for the dataset using nested <details> elements, ordered and labeled. */
-function getDatasetHtml(webview: vscode.Webview, dataset: NetCDFDataset, fileName: string = 'NetCDF File'): string {
-  const alwaysExpandable = new Set(['dtype', 'shape', 'dims', 'encoding']);
-  function getSampleSlice(shape: number[] | undefined, sampleCount: number = 10): string {
-    if (!shape || shape.length === 0) {
-      return '';
-    }
-    let remaining = sampleCount;
-    const slices = [];
-    for (let i = 0; i < shape.length; ++i) {
-      if (remaining <= 0) {
-        slices.push('0:1');
-        continue;
-      }
-      let thisDim = 1;
-      for (let j = i + 1; j < shape.length; ++j) {
-        thisDim *= shape[j];
-      }
-      let count = Math.min(shape[i], Math.ceil(remaining / thisDim));
-      slices.push(`0:${count}`);
-      remaining = Math.floor(remaining / count);
-    }
-    return `[${slices.join(', ')}]`;
-  }
-
-  function renderTree(node: unknown, label: string, indent = 0, parent?: Record<string, any>): string {
-    // Special handling for sample_data
-    let displayLabel = label;
-    if (label === 'sample_data' && Array.isArray(node) && parent && (parent.shape || parent.dims)) {
-      const shape: number[] =
-        parent.shape || (parent.dims ? parent.dims.map((d: string) => parent[d]?.length || 0) : []);
-      const sampleSlice = getSampleSlice(shape, node.length);
-      displayLabel = `sample_data ${sampleSlice}`;
-    }
-
-    // Always expandable for certain keys, even if primitive
-    if (alwaysExpandable.has(label)) {
-      let content = '';
-      if (typeof node === 'object' && node !== null) {
-        content = Object.entries(node)
-          .map(([k, v]) => renderTree(v, k, indent + 1, node))
-          .join('');
-      } else {
-        content = `<div style="padding-left:${(indent + 1) * 20}px"><span class="val">${node}</span></div>`;
-      }
-      return `<details>
-        <summary style="padding-left:${indent * 20}px">${displayLabel}</summary>
-        ${content}
-      </details>`;
-    }
-
-    // For plain objects, render as expandable
-    if (typeof node === 'object' && node !== null && !Array.isArray(node)) {
-      const children = Object.entries(node)
-        .map(([k, v]) => renderTree(v, k, indent + 1, node))
-        .join('');
-      return `<details>
-      <summary style="padding-left:${indent * 20}px">${displayLabel}</summary>
-      ${children}
-    </details>`;
-    }
-
-    // For arrays, show a preview (first 10 values)
-    if (Array.isArray(node)) {
-      const preview = node
-        .slice(0, 10)
-        .map((v, i) => `<div style="padding-left:${(indent + 1) * 20}px">[${i}]: <span class="val">${v}</span></div>`)
-        .join('');
-      return `<details>
-      <summary style="padding-left:${indent * 20}px">${displayLabel}</summary>
-      ${preview}
-    </details>`;
-    }
-
-    // For primitives, just show as a line
-    return `<div style="padding-left:${indent * 20}px">${displayLabel}: <span class="val">${node}</span></div>`;
-  }
-
-  // Prepare ordered branches
-  const dims = dataset.dims ? renderTree(dataset.dims, 'Dimensions', 1) : '';
-  const coords = dataset.coords ? renderTree(dataset.coords, 'Coordinates', 1) : '';
-  const dataVars = dataset.data_vars ? renderTree(dataset.data_vars, 'Data Variables', 1) : '';
-  const globalAttrs = dataset.attrs ? renderTree(dataset.attrs, 'Global Attributes', 1) : '';
-
-  return `
-  <!DOCTYPE html>
-  <html>
-  <head>
-    <style>
-      body { font-family: sans-serif; padding: 16px; }
-      summary { font-weight: bold; cursor: pointer; }
-      details { margin-bottom: 8px; }
-      .val { color: #333; }
-    </style>
-  </head>
-  <body>
-    <h1>🔍 NetCDF Structure Viewer</h1>
-    <details>
-      <summary style="font-size:1.2em;">${fileName}</summary>
-      ${dims}
-      ${coords}
-      ${dataVars}
-      ${globalAttrs}
-    </details>
-  </body>
-  </html>
-  `;
-}
-
-/**
- * Generates HTML for the Webview, including metadata and a mini-chart.
- */
-function getWebviewContent(webview: vscode.Webview, context: vscode.ExtensionContext, variable: NamedVariable): string {
-  // URI for local Chart.js script in media folder
-  const chartJsUri = webview.asWebviewUri(vscode.Uri.joinPath(context.extensionUri, 'media', 'chart.js'));
-
-  // Format variable summary like xarray
-  const dimsStr =
-    variable.dims && variable.dims.length > 0
-      ? `(${variable.dims
-          .map((d: string, i: number) => `${d}: ${variable.shape ? variable.shape[i] : '?'}`)
-          .join(', ')})`
-      : '';
-
-  const attrsStr = Object.entries(variable.attrs || {})
-    .map(([k, v]) => `<tr><td>${k}</td><td>${JSON.stringify(v)}</td></tr>`)
-    .join('');
-
-  // Read all data and take the first 10 values as sample
-  let rawData: any[] = [];
-  if (Array.isArray(variable.sample_data)) {
-    rawData = variable.sample_data;
-  }
-  const sampleData = rawData.slice(0, 10);
-
-  // Add the sample slice function
-  function getSampleSlice(shape: number[] | undefined, sampleCount: number = 10): string {
-    if (!shape || shape.length === 0) {
-      return '';
-    }
-    let remaining = sampleCount;
-    const slices = [];
-    for (let i = 0; i < shape.length; ++i) {
-      if (remaining <= 0) {
-        slices.push('0:1');
-        continue;
-      }
-      let thisDim = 1;
-      for (let j = i + 1; j < shape.length; ++j) {
-        thisDim *= shape[j];
-      }
-      let count = Math.min(shape[i], Math.ceil(remaining / thisDim));
-      slices.push(`0:${count}`);
-      remaining = Math.floor(remaining / count);
-    }
-    return `[${slices.join(', ')}]`;
-  }
-  const shape = variable.shape || (variable.dims ? variable.dims.map((d: string) => variable[d]?.length || 0) : []);
-  const sampleSlice = getSampleSlice(shape, sampleData.length);
-
-  return `<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta http-equiv="Content-Security-Policy"
-        content="default-src 'none'; img-src ${
-          webview.cspSource
-        } https:; script-src 'nonce-chart' ${webview.cspSource}; style-src ${webview.cspSource};">
-    <style>
-        body { font-family: sans-serif; padding: 16px; }
-        table { border-collapse: collapse; width: 100%; margin-bottom: 16px; }
-        th, td { border: 1px solid #ccc; padding: 4px 8px; text-align: left; }
-    </style>
-    <title>Variable Preview</title>
-</head>
-<body>
-    <h1>${variable.name || variable.label || '?'}</h1>
-    <p><strong>Dimensions:</strong> ${
-      variable.dims && variable.shape
-        ? variable.dims.map((d: string, i: number) => `${d} (${variable.shape[i]})`).join(' × ')
-        : (variable.dims || []).join(' × ')
-    }</p>
-    <p><strong>Type:</strong> ${variable.dtype || '?'}</p>
-
-    <h2>Attributes</h2>
-    <table>
-        <tr><th>Key</th><th>Value</th></tr>
-        ${attrsStr || '<tr><td colspan="2"><em>No attributes</em></td></tr>'}
-    </table>
-
-    <h2>
-      Sample Data${sampleSlice ? ` ${sampleSlice}` : ''} (first ${sampleData.length} values)
-    </h2>
-${
-  sampleData.length > 0
-    ? `
-      <table>
-        <tr><th>Index</th><th>Value</th></tr>
-        ${sampleData.map((v, i) => `<tr><td>${i}</td><td>${v}</td></tr>`).join('')}
-      </table>
-      <canvas id="chart" width="400" height="200"></canvas>
-      `
-    : '<p><em>No data available for this variable.</em></p>'
-}
-    ${
-      sampleData.length > 0
-        ? `
-    <!-- Acquire VS Code API -->
-    <script nonce="nonce-chart">
-        const vscode = acquireVsCodeApi();
-    </script>
-    <!-- Chart.js library -->
-    <script nonce="nonce-chart" src="${chartJsUri}"></script>
-    <!-- Render the chart -->
-    <script nonce="nonce-chart">
-        const ctx = document.getElementById('chart').getContext('2d');
-        new Chart(ctx, {
-            type: 'line',
-            data: {
-                labels: ${JSON.stringify(sampleData.map((_, i) => i))},
-                datasets: [{
-                    label: '${variable.name || variable.label || '?'}',
-                    data: ${JSON.stringify(sampleData)},
-                    fill: false,
-                    tension: 0.1
-                }]
-            },
-            options: { responsive: true }
-        });
-
-        ctx.canvas.addEventListener('click', () => {
-            vscode.postMessage({ command: 'alert', text: 'Chart clicked!' });
-        });
-    </script>
-    `
-        : ''
-    }
-</body>
-</html>`;
-}
-
-/** Data attached to tree items - either the full dataset or a named variable */
-type TreeItemData = NetCDFDataset | NamedVariable | undefined;
-
-/**
- * Represents an item in the NetCDF Explorer view.
- */
-class NetCDFTreeItem extends vscode.TreeItem {
-  constructor(
-    public readonly label: string,
-    public readonly state: vscode.TreeItemCollapsibleState,
-    public readonly data: TreeItemData = undefined,
-    public readonly contextValue: string = ''
-  ) {
-    super(label, state);
-    this.contextValue = contextValue;
-  }
-}
-
-/**
- * Provides data (variables) for the NetCDF Explorer tree view.
- */
-export class NetCDFTreeProvider implements vscode.TreeDataProvider<NetCDFTreeItem> {
-  private _onDidChange = new vscode.EventEmitter<NetCDFTreeItem | undefined>();
-  readonly onDidChangeTreeData = this._onDidChange.event;
-
-  constructor(private context: vscode.ExtensionContext) {}
-
-  /** Refresh the entire tree */
-  refresh(): void {
-    this._onDidChange.fire(undefined);
-  }
-
-  getTreeItem(item: NetCDFTreeItem): vscode.TreeItem {
-    return item;
-  }
-
-  async getChildren(element?: NetCDFTreeItem): Promise<NetCDFTreeItem[]> {
-    const stored = this.context.workspaceState.get<StoredNetCDF>('lastNetCDF');
-    if (!stored?.dataset) {
-      return [];
-    }
-    const ds: NetCDFDataset = stored.dataset;
-    const fileName = stored.uri ? path.basename(stored.uri.fsPath) : 'NetCDF File';
-    // If no element, show the file name as the root
-    if (!element) {
-      return [new NetCDFTreeItem(fileName, vscode.TreeItemCollapsibleState.Expanded, ds, 'file')];
-    }
-
-    // If the element is the file node, show the main branches
-    if (element.contextValue === 'file') {
-      return [
-        new NetCDFTreeItem('Dimensions', vscode.TreeItemCollapsibleState.Collapsed),
-        new NetCDFTreeItem('Coordinates', vscode.TreeItemCollapsibleState.Collapsed),
-        new NetCDFTreeItem('Data Variables', vscode.TreeItemCollapsibleState.Collapsed),
-      ];
-    }
-
-    if (element.label === 'Dimensions') {
-      return Object.entries(ds.dims || {}).map(
-        ([dim, size]) => new NetCDFTreeItem(`${dim} (${size})`, vscode.TreeItemCollapsibleState.None)
-      );
-    }
-
-    if (element.label === 'Coordinates') {
-      return Object.entries(ds.coords || {}).map(([coordName, v]) => {
-        const namedVar: NamedVariable = { ...v, name: coordName };
-        return new NetCDFTreeItem(
-          coordName,
-          vscode.TreeItemCollapsibleState.Collapsed,
-          namedVar,
-          'variable'
-        );
-      });
-    }
-
-    if (element.label === 'Data Variables') {
-      return Object.entries(ds.data_vars || {}).map(([varName, v]) => {
-        const namedVar: NamedVariable = { ...v, name: varName };
-        return new NetCDFTreeItem(
-          varName,
-          vscode.TreeItemCollapsibleState.Collapsed,
-          namedVar,
-          'variable'
-        );
-      });
-    }
-
-    // If this is a variable, show its attributes, sample data, and encoding as children
-    if (element.contextValue === 'variable' && element.data) {
-      const varData = element.data as NamedVariable;
-      return [
-        new NetCDFTreeItem('Attributes', vscode.TreeItemCollapsibleState.Collapsed, varData, 'attributes'),
-        new NetCDFTreeItem('Sample Data', vscode.TreeItemCollapsibleState.Collapsed, varData, 'sample'),
-        new NetCDFTreeItem('Encoding', vscode.TreeItemCollapsibleState.Collapsed, varData, 'encoding'),
-      ];
-    }
-
-    // Show encoding children
-    if (element.contextValue === 'encoding' && element.data) {
-      const varData = element.data as NamedVariable;
-      return Object.entries(varData.encoding || {}).map(
-        ([k, v]) => new NetCDFTreeItem(`${k}: ${JSON.stringify(v)}`, vscode.TreeItemCollapsibleState.None)
-      );
-    }
-
-    // Show attribute children
-    if (element.contextValue === 'attributes' && element.data) {
-      const varData = element.data as NamedVariable;
-      return Object.entries(varData.attrs || {}).map(
-        ([k, v]) => new NetCDFTreeItem(`${k}: ${JSON.stringify(v)}`, vscode.TreeItemCollapsibleState.None)
-      );
-    }
-
-    // Show sample data children
-    if (element.contextValue === 'sample' && element.data) {
-      const varData = element.data as NamedVariable;
-      const sampleData = Array.isArray(varData.sample_data) ? varData.sample_data.slice(0, 10) : [];
-      return sampleData.map(
-        (v, i) => new NetCDFTreeItem(`[${i}]: ${v}`, vscode.TreeItemCollapsibleState.None)
-      );
-    }
-
-    return [];
-  }
-}
-
-/**
- * Inspects a NetCDF file using an external Python script and returns the parsed output.
- */
-async function inspectNetCDFWithPython(context: vscode.ExtensionContext, filePath: string): Promise<InspectResult> {
-  const scriptPath = path.join(context.extensionPath, 'inspect_netcdf.py');
-  const pythonPath = vscode.workspace.getConfiguration().get<string>('netcdfViewer.pythonPath', 'python');
-  return new Promise((resolve, reject) => {
-    execFile(pythonPath, [scriptPath, filePath], { maxBuffer: 10 * 1024 * 1024 }, (err, stdout, stderr) => {
-      if (err) {
-        // Show stderr in the error message for debugging
-        reject(`Python error: ${stderr || err.message}`);
-      } else {
-        try {
-          resolve(JSON.parse(stdout));
-        } catch (e) {
-          reject('Failed to parse Python output: ' + stdout);
-        }
-      }
-    });
-  });
-}
-
-/**
- * Checks Python and required dependencies, returning an error message if any are missing.
- */
-async function checkPythonDependencies(pythonPath: string): Promise<string | null> {
-  return new Promise((resolve) => {
-    execFile(pythonPath, ['-c', 'import xarray; import netCDF4'], (err) => {
-      if (err) {
-        resolve('Python, xarray, or netCDF4 not found. Please check your Python path and environment.');
-      } else {
-        resolve(null);
-      }
-    });
-  });
 }
