+++ src/views/variableWebview.ts
@@ -1,0 +1,141 @@
+import * as vscode from 'vscode';
+import { NamedVariable } from '../types';
+import { getSampleSlice } from '../utils/sampleSlice';
+
+/**
+ * Opens a Webview panel to preview the selected NetCDF variable.
+ *
+ * @param context - VS Code extension context
+ * @param variable - The variable to preview
+ */
+export function showVariableWebview(context: vscode.ExtensionContext, variable: NamedVariable): void {
+  const panel = vscode.window.createWebviewPanel(
+    'netcdfVarPreview',
+    `Preview: ${variable.name || variable.label || '?'}`,
+    vscode.ViewColumn.One,
+    {
+      enableScripts: true,
+      localResourceRoots: [vscode.Uri.joinPath(context.extensionUri, 'media')],
+    }
+  );
+
+  panel.webview.html = getWebviewContent(panel.webview, context, variable);
+
+  panel.webview.onDidReceiveMessage((msg) => {
+    if (msg.command === 'alert') {
+      vscode.window.showInformationMessage(msg.text);
+    }
+  });
+}
+
+/**
+ * Generates HTML for the Webview, including metadata and a mini-chart.
+ *
+ * @param webview - The webview instance
+ * @param context - VS Code extension context
+ * @param variable - The variable to display
+ * @returns HTML string
+ */
+export function getWebviewContent(
+  webview: vscode.Webview,
+  context: vscode.ExtensionContext,
+  variable: NamedVariable
+): string {
+  // URI for local Chart.js script in media folder
+  const chartJsUri = webview.asWebviewUri(vscode.Uri.joinPath(context.extensionUri, 'media', 'chart.js'));
+
+  // Format variable summary like xarray
+  const dimsStr =
+    variable.dims && variable.dims.length > 0
+      ? `(${variable.dims.map((d: string, i: number) => `${d}: ${variable.shape ? variable.shape[i] : '?'}`).join(', ')})`
+      : '';
+
+  const attrsStr = Object.entries(variable.attrs || {})
+    .map(([k, v]) => `<tr><td>${k}</td><td>${JSON.stringify(v)}</td></tr>`)
+    .join('');
+
+  // Read all data and take the first 10 values as sample
+  const rawData: (number | string | null)[] = Array.isArray(variable.sample_data) ? variable.sample_data : [];
+  const sampleData = rawData.slice(0, 10);
+
+  const shape = variable.shape || [];
+  const sampleSlice = getSampleSlice(shape, sampleData.length);
+
+  return `<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta http-equiv="Content-Security-Policy"
+        content="default-src 'none'; img-src ${webview.cspSource} https:; script-src 'nonce-chart' ${webview.cspSource}; style-src ${webview.cspSource};">
+    <style>
+        body { font-family: sans-serif; padding: 16px; }
+        table { border-collapse: collapse; width: 100%; margin-bottom: 16px; }
+        th, td { border: 1px solid #ccc; padding: 4px 8px; text-align: left; }
+    </style>
+    <title>Variable Preview</title>
+</head>
+<body>
+    <h1>${variable.name || variable.label || '?'}</h1>
+    <p><strong>Dimensions:</strong> ${
+      variable.dims && variable.shape
+        ? variable.dims.map((d: string, i: number) => `${d} (${variable.shape[i]})`).join(' × ')
+        : (variable.dims || []).join(' × ')
+    }</p>
+    <p><strong>Type:</strong> ${variable.dtype || '?'}</p>
+
+    <h2>Attributes</h2>
+    <table>
+        <tr><th>Key</th><th>Value</th></tr>
+        ${attrsStr || '<tr><td colspan="2"><em>No attributes</em></td></tr>'}
+    </table>
+
+    <h2>
+      Sample Data${sampleSlice ? ` ${sampleSlice}` : ''} (first ${sampleData.length} values)
+    </h2>
+${
+  sampleData.length > 0
+    ? `
+      <table>
+        <tr><th>Index</th><th>Value</th></tr>
+        ${sampleData.map((v, i) => `<tr><td>${i}</td><td>${v}</td></tr>`).join('')}
+      </table>
+      <canvas id="chart" width="400" height="200"></canvas>
+      `
+    : '<p><em>No data available for this variable.</em></p>'
+}
+    ${
+      sampleData.length > 0
+        ? `
+    <!-- Acquire VS Code API -->
+    <script nonce="nonce-chart">
+        const vscode = acquireVsCodeApi();
+    </script>
+    <!-- Chart.js library -->
+    <script nonce="nonce-chart" src="${chartJsUri}"></script>
+    <!-- Render the chart -->
+    <script nonce="nonce-chart">
+        const ctx = document.getElementById('chart').getContext('2d');
+        new Chart(ctx, {
+            type: 'line',
+            data: {
+                labels: ${JSON.stringify(sampleData.map((_, i) => i))},
+                datasets: [{
+                    label: '${variable.name || variable.label || '?'}',
+                    data: ${JSON.stringify(sampleData)},
+                    fill: false,
+                    tension: 0.1
+                }]
+            },
+            options: { responsive: true }
+        });
+
+        ctx.canvas.addEventListener('click', () => {
+            vscode.postMessage({ command: 'alert', text: 'Chart clicked!' });
+        });
+    </script>
+    `
+        : ''
+    }
+</body>
+</html>`;
+}
